[cpu_backend] Refine x86 FP16 GEMM/GEMV: namespace, portability, and tests
- Restore the column-major FP16 sgemm path: row-major uses the cache-blocked
  hgemm, while column-major uses the FP32-conversion + CBLAS path because
  __fallback_sgemm ignores the storage order.
- Make the hgemm aligned allocator portable to non-POSIX targets, using
  _aligned_malloc / _aligned_free on Windows and std::aligned_alloc /
  std::free elsewhere, with the matching free path branched.
- Document the cache-blocking constants in hgemm_common.h with the cache
  tier and byte budget each value targets.
- Compile the workspace reallocation counters out of non-ENABLE_TEST builds.
- Track every thread-local workspace in a registry and release it through
  release_all_hgemm_workspaces() at teardown so the reused pool buffers are
  freed deterministically.
- Move the x86 hgemm tree from nntrainer::avx2 into a dedicated
  nntrainer::hgemm namespace.
- Keep the internal x86 hgemm headers off the install manifest (include-path
  only) so the Tizen and Debian packaging stays consistent.
- Split the x86 GEMM/GEMV unit tests into a new x86-only
  unittest_nntrainer_cpu_backend_fp16_x86.cpp and restore the Android-only
  unittest_nntrainer_cpu_backend_fp16.cpp to its original contents.
- Apply clang-format-14 to the changed lines.
