+++ nntrainer/tensor/cpu_backend/x86/hgemm/hgemm_kernel/hgemm_kernel_1x8.cpp
@@ -14,7 +14,7 @@
 
 #include <immintrin.h>
 
-namespace nntrainer::avx2::internal {
+namespace nntrainer::hgemm::internal {
 
 void hgemm_kernel_1x8(unsigned int K, const float *packed_A,
                       const float *packed_B, float *C, unsigned int c_stride) {
@@ -34,4 +34,4 @@ void hgemm_kernel_1x8(unsigned int K, const float *packed_A,
   _mm256_storeu_ps(C, c00);
 }
 
-} /* namespace nntrainer::avx2::internal */
+} /* namespace nntrainer::hgemm::internal */
