+++ nntrainer/tensor/cpu_backend/x86/hgemm/hgemm_kernel/meson.build
@@ -1,7 +1,3 @@
-x86_hgemm_kernel_headers = [
-  'hgemm_kernel.h',
-]
-
 x86_hgemm_kernel_sources = [
   'hgemm_kernel_1x8.cpp',
   'hgemm_kernel_1x16.cpp',
@@ -16,8 +12,4 @@ x86_hgemm_kernel_sources = [
 
 foreach s : x86_hgemm_kernel_sources
   nntrainer_sources += meson.current_source_dir() / s
-endforeach
-
-foreach h : x86_hgemm_kernel_headers
-  nntrainer_headers += meson.current_source_dir() / h
 endforeach
